removed useless code
